DogBreedApp2/requirements.txt
Flask==1.0.2
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
click==6.7
cycler==0.10.0
gast==0.2.0
grpcio==1.12.1
h5py==2.8.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.1.0
pip==9.0.3
protobuf==3.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==4.2b4
scikit-learn==0.19.1
scipy==1.1.0
setuptools==39.2.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
tqdm==4.23.4
Werkzeug==0.14.1
wheel==0.31.1
gunicorn==19.9.0
